[fix] Potential issue

+++ reloadAnyResponse.php
@@ -5,7 +5,7 @@
  * @author Denis Chenu <denis@sondages.pro>
  * @copyright 2018-2020 Denis Chenu <http://www.sondages.pro>
  * @license AGPL v3
- * @version 1.5.0
+ * @version 1.5.1
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -585,7 +585,9 @@ class reloadAnyResponse extends PluginBase {
     {
         $surveyId = $this->getEvent()->get('surveyId');
         $responseId = $this->getEvent()->get('responseId');
-        \reloadAnyResponse\models\surveySession::model()->deleteByPk(array('sid'=>$surveyId,'srid'=>$responseId));
+        if(!empty($surveyId) && !empty($responseId)) {
+            \reloadAnyResponse\models\surveySession::model()->deleteByPk(array('sid'=>$surveyId,'srid'=>$responseId));
+        }
     }
 
     /** @inheritdoc **/
